refactor(browser): split views_browser into places, audition, remote, style
