+++ crates/vibez-ui/src/app/mod.rs
@@ -137,6 +137,10 @@ mod plugins;
 mod project_io;
 mod update;
 mod views_browser;
+mod views_browser_audition;
+mod views_browser_places;
+mod views_browser_remote;
+mod views_browser_style;
 mod views_detail;
 mod views_devices;
 mod views_overlays;
